fix: attempt 1 of fixing retry errors comming from skelet.py
